+++ apps/web/src/layouts/auth-layout.tsx
@@ -2,17 +2,18 @@ import { Navigate, Outlet, useLocation } from 'react-router-dom';
 import { SessionLoadingScreen } from '../screen/session-loading-screen';
 import { AccountProvider, useSession } from '@cbdc-markka/utils-react';
 
+const loginRedirectExemptions = ['/auth/logout'];
 export function AuthLayout() {
   const { status, session } = useSession();
   const location = useLocation();
   //const status: any = 'loading';
 
-  if (status === 'unauthenticated') {
+  if (status === 'unauthenticated' && !loginRedirectExemptions.includes(location.pathname)) {
     const url = `/login?callback_url=${location.pathname}`;
     return <Navigate to={url} />;
   } else if (status === 'loading') {
     return <SessionLoadingScreen />;
-  } else if (session.user.status === 'pending') {
+  } else if (session?.user.status === 'pending') {
     return <Navigate to='/pending-user' />;
   }