+++ apps/web/src/screen/verify-email-screen.tsx
@@ -5,24 +5,23 @@ import { useStatus } from '../hooks/use-status';
 import { useEffect, useRef } from 'react';
 import { NoticeScreen } from '../components/notice-screen';
 import { Spinner, SpinnerTimer } from '../components/spinner';
-import { Button } from '../components/button';
+import { Button, LoaderButton } from '../components/button';
+import { ErrorMessage, SuccessMessage } from '../components/helper-message';
 
+/**Verifies a users email in the background and redirects to the login page if it succeeds. */
 export function VerifyEmailScreen() {
   const navigate = useNavigate();
   const { apiInterface } = useApi();
   const [searchParams] = useSearchParams();
   const { status, setStatus, loading } = useStatus();
+  const success = status === 'success';
   const token = searchParams.get('token');
 
   const handleVerification = async () => {
     try {
       setStatus('loading');
-
       const res = await apiInterface.verifyUserById({ token });
       if (!res.ok) {
-        if (res.status === 409) {
-          navigate('/');
-        }
         setStatus('error');
       } else {
         setStatus('success');
@@ -51,17 +50,32 @@ export function VerifyEmailScreen() {
   return (
     <NoticeScreen
       title='Vahvistetaan sähköpostiosoitetta'
-      bodyText='Ole hyvä ja odota kun vahvistamme sähköpostiosoitteesi...'>
-      {status === 'loading' ? (
-        <Spinner />
-      ) : status !== 'idle' && status !== 'loading' && status !== 'success' ? (
-        <Button
-          onClick={handleVerification}
-          type='button'
-          rounded>
-          Yritä uudelleen
-        </Button>
-      ) : null}
-    </NoticeScreen>
+      bodyText='Ole hyvä ja odota kun vahvistamme sähköpostiosoitteesi...'
+      footer={
+        <div className='flex flex-col gap-4 w-full'>
+          <div className='flex flex-col gap-2 w-full'>
+            <LoaderButton
+              loading={loading}
+              disabled={loading || success}
+              onClick={handleVerification}
+              type='button'
+              rounded>
+              Yritä uudelleen
+            </LoaderButton>
+            <Button
+              onClick={() => navigate('/')}
+              disabled={loading || success}
+              variant='outlined'
+              rounded>
+              Peruuta
+            </Button>
+          </div>
+          {status !== 'loading' && status !== 'idle' && status !== 'success' ? (
+            <ErrorMessage>Jotain meni pieleen!</ErrorMessage>
+          ) : status === 'success' ? (
+            <SuccessMessage>Sähköpostin vahvistus onnistui!</SuccessMessage>
+          ) : null}
+        </div>
+      }></NoticeScreen>
   );
 }